requirements.txt
psycopg2-binary>=2.9
python-dotenv>=1.0
pandas>=2.2
numpy>=1.26
requests>=2.31
yfinance>=0.2.40
pandas-market-calendars>=4.4
tqdm>=4.66
